backend/requirements.txt
anyio==3.6.1
fastapi==0.78.0
idna==3.3
joblib==1.1.0
numpy==1.22.4
Pillow==9.1.1
pydantic==1.9.1
scikit-learn==1.1.1
scipy==1.8.1
sklearn==0.0
sniffio==1.2.0
starlette==0.19.1
threadpoolctl==3.1.0
typing-extensions==4.2.0
